apps/api/requirements.txt
fastapi
uvicorn
mangum
boto3
email-validator
passlib[bcrypt]
PyJWT
